# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music Studio Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List all songs
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a song (link lyric + instrumental)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSong"
      responses:
        "201":
          description: Created song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/{songId}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "200":
          description: Song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found
    put:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      parameters:
        - $ref: "#/components/parameters/songId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSong"
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - $ref: "#/components/parameters/songId"
      responses:
        "204":
          description: Deleted

  # ─── Releases ────────────────────────────────────────────────────────────────
...
components:
  parameters:
...
    songId:
      name: songId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Song:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
          nullable: true
        lyricId:
          type: integer
          nullable: true
        instrumentalId:
          type: integer
          nullable: true
        title:
          type: string
        artistName:
          type: string
          nullable: true
        featuredArtists:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        isrc:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        copyright:
          type: string
          nullable: true
        recordLabel:
          type: string
          nullable: true
        status:
          type: string
          enum: [idea, writing, recording, mixing, mastering, ready, released]
        coverImageUrl:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, createdAt, updatedAt]
    CreateSong:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        lyricId:
          type: integer
          nullable: true
        instrumentalId:
          type: integer
          nullable: true
        title:
          type: string
        artistName:
          type: string
          nullable: true
        featuredArtists:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        isrc:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        copyright:
          type: string
          nullable: true
        recordLabel:
          type: string
          nullable: true
        status:
          type: string
          enum: [idea, writing, recording, mixing, mastering, ready, released]
          default: idea
        coverImageUrl:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title]
    UpdateSong:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        lyricId:
          type: integer
          nullable: true
        instrumentalId:
          type: integer
          nullable: true
        title:
          type: string
        artistName:
          type: string
          nullable: true
        featuredArtists:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        isrc:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        copyright:
          type: string
          nullable: true
        recordLabel:
          type: string
          nullable: true
        status:
          type: string
          enum: [idea, writing, recording, mixing, mastering, ready, released]
        coverImageUrl:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ─── Release ───────────────────────────────────────────────────────────────
...