# lib/api-zod/src/generated/types/createSong.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSongStatus } from "./createSongStatus";

export interface CreateSong {
  projectId?: number | null;
  lyricId?: number | null;
  instrumentalId?: number | null;
  title: string;
  artistName?: string | null;
  featuredArtists?: string | null;
  genre?: string | null;
  mood?: string | null;
  bpm?: number | null;
  musicalKey?: string | null;
  durationSeconds?: number | null;
  isrc?: string | null;
  upc?: string | null;
  copyright?: string | null;
  recordLabel?: string | null;
  status?: CreateSongStatus;
  coverImageUrl?: string | null;
  fileUrl?: string | null;
  notes?: string | null;
}